Safari-611.1.10/LayoutTests/fast/canvas/canvas-measureText-ideographicSpace-expected.txt

PASS spanWidth is canvasWidth
PASS successfullyParsed is true

TEST COMPLETE
